% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsingModeGivenS.R
\name{dsingModeGivenS}
\alias{dsingModeGivenS}
\title{Probability of a single mode}
\usage{
dsingModeGivenS(S, mod, pars, modeCutOff = 0.1, relative = TRUE,
  B = 10000)
}
\arguments{
\item{S}{number of species}

\item{mod}{the SAD model specified by the \code{d*} function, e.g. \code{dfish}}

\item{pars}{model parameters}

\item{modeCutOff}{the fraction by which a peak must exceed its neigbors to count as a mode}

\item{relative}{logical, should \code{modeCutOff} be taken as relative or not}

\item{B}{number of Monte Carlo iterations}
}
\value{
a numeric giving the desired probability
}
\description{
For a given model and number of species, returns the probability of a
single mode in the Preston plot
}
\details{
A mode is defined by the cuttoff \code{modeCutOff}. This can be made a proportional cuttoff if \code{relative = TRUE} is relative
to the height of the peak. For example if \code{modeCutOff = 0.1, relative = TRUE} this sequence would contain a mode:
\code{c(17, 20, 17)} but this one would not: \code{c(19, 20, 19)}.  The Probability of a single mode
is numerically involved and so a Monte Carlo approximation is used with \code{B} replicates.
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
