% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expAxis.R
\name{expAxis}
\alias{expAxis}
\title{Exponential axis}
\usage{
expAxis(side, at, ...)
}
\arguments{
\item{side}{side to add axis to}

\item{at}{locations of ticks}

\item{...}{additional parameters passed to \code{axis}}
}
\description{
When plotting already log-transformed data, add an axis with exponential
tick labels
}
\details{
assumes \eqn{log_{10}}
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
