% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectedN.R
\name{expectedN}
\alias{expectedN}
\alias{solve4expectedN}
\title{Expected N given S and an SAD model}
\usage{
expectedN(S, mod, pars)

solve4expectedN(N, mod, pars)
}
\arguments{
\item{S}{number of spcies}

\item{mod}{string specifying the SAD model (e.g. \code{'fish'})}

\item{pars}{vector of parameters for the SAD model}

\item{N}{total number of individuals}
}
\value{
\code{expectedN} returns the expectation of N, while \code{solve4expectedN}
finds the value of S that gives the desired expected value of N given the SAD model
}
\description{
The expected value of N, the sum of all abundances over all species
}
\details{
Uses the \code{dNGivenS} function internally.
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
