% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localMaxima.R
\name{localMaxima}
\alias{localMaxima}
\title{Find local maxima}
\usage{
localMaxima(x, by = 1, xmin = 0)
}
\arguments{
\item{x}{vector of sequential numbers within which to look for local maxima}

\item{by}{minimum value by which a maximum must exceed its neighbors}

\item{xmin}{the hypothetical minimum value \code{x} might achieve}
}
\value{
a vector of indeces of the local maxima
}
\description{
Returns a vector of indeces of the local maxima
}
\details{
Thanks to stackoverflow user stas g for the starting point solution in
https://stackoverflow.com/questions/34205515/finding-local-maxima-and-minima-in-r
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
