% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSGivenN.R
\name{dNGivenS}
\alias{dNGivenS}
\alias{pNGivenS}
\alias{qNGivenS}
\title{Probability of N given S}
\usage{
dNGivenS(N, S, pars, mod, log = FALSE)

pNGivenS(N, S, pars, mod, lower.tail = TRUE, log.p = FALSE)

qNGivenS(p, S, pars, mod)
}
\arguments{
\item{N}{total abundance of interest}

\item{S}{number of species}

\item{pars}{parameter(s) for the SAD model}

\item{mod}{a string specifying the SAD model}

\item{log}{logical, should the log probability be used}

\item{lower.tail}{logical, should the lower tail be used}

\item{log.p}{logical, should the log cumulative probability be used}

\item{p}{probability for which to return the quantile}
}
\value{
A numeric vector of length equal to the input
}
\description{
Density, distribution, and quantile functions for the total number of
individuals given the number of species
}
\details{
These functions assume a sufficiently large number of species (in practice >= 20)
such that the Central Limit Theorem applies and we can approxiate the distribution of the
sum of \code{S} random variables as a normal distribution
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
