% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmultinom.R
\name{pmultinom}
\alias{pmultinom}
\title{Multinomial CDF}
\usage{
pmultinom(q, size, prob, s = size)
}
\arguments{
\item{q}{vector of quantiles, k-long}

\item{size}{total number of objects put into k boxes}

\item{prob}{vector of probabilities for each of k boxes, internally normalized}
}
\value{
a vector of probabilities equal in length to \code{q}
}
\description{
Cumulative probability function of the multinomial distribution
}
\references{
Levin, B. (1981). A representation for multinomial cumulative distribution functions.
The Annals of Statistics, 1123-1126.
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
