% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prestonize.R
\name{prestonize}
\alias{prestonize}
\alias{prestonize.integer}
\alias{prestonize.sad}
\title{'Prestonize' an SAD}
\usage{
prestonize(x)

\method{prestonize}{integer}(x)

\method{prestonize}{sad}(x)
}
\arguments{
\item{x}{object to Prestonize}
}
\value{
an object of class \code{preston} showing the observed and theoretical
probabilities for each Preston octave, and recording the total number of species
}
\description{
Bin an SAD into Preston octaves
}
\details{
Works for both theoretical SADs (i.e. the density function, by summing
probabilty mass) and for empirical SADs (i.e. by binning abundances and summing
frequencies)
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
