% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sadShape.R
\name{sadShape}
\alias{sadShape}
\title{Plotting SAD shapes}
\usage{
sadShape(pars, nn, dfun, main, ...)
}
\arguments{
\item{pars}{a list containing the parameter values of interest}

\item{nn}{vector of integer abundances for which to plot probabilities}

\item{dfun}{the density function of interest}

\item{main}{plot title}

\item{...}{additional parameters passed to \code{plot}}
}
\description{
plot the shapes of SADs in parameter space
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
