% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbimage.R
\name{sbimage}
\alias{sbimage}
\title{Species by beta image}
\usage{
sbimage(s, b, p, colz, parArgs, legendLab, ...)
}
\arguments{
\item{s}{the species values used to calculate \code{p}}

\item{b}{the beta values used to calculate \code{p}}

\item{p}{the matrix of probabilities (rows correspond to species values)}

\item{colz}{vector of colors to use}

\item{legendLab}{lable for the legend}

\item{...}{additional parameters passed to \code{image}}

\item{list}{of arguments to pass to \code{par}, note, at a minimum must include \code{mar}}
}
\description{
An image plot customized for a matrix of probabilities across different
numbers of species and beta parameters
}
\details{
assumes \code{b} was made over a \eqn{log_{10}} scale
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
