% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGivenN.R
\name{dSGivenN}
\alias{dSGivenN}
\title{Probability of S given N}
\usage{
dSGivenN(S, N, pars, mod)
}
\arguments{
\item{S}{number of species}

\item{N}{total abundance}

\item{pars}{parameter(s) for the SAD model}

\item{mod}{a string specifying the SAD model}
}
\value{
A numeric vector of length equal to the input
}
\description{
Density function for the number of species given the total number of sampled individuals
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
