% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmultinom.R
\name{dsumuptpois}
\alias{dsumuptpois}
\title{Sum of upper truncated Poisson random variables}
\usage{
dsumuptpois(N, lambda, m)
}
\arguments{
\item{N}{integer, the value of summed upper truncated Poisson r.v.'s}

\item{lambda}{vector of means for each upper truncted Poisson}

\item{m}{integer vector upper (inclusive) limits}
}
\description{
Sum of upper truncated Poisson random variables
}
\references{
Levin, B. (1981). A representation for multinomial cumulative distribution functions. The Annals of Statistics, 1123-1126.
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
