% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmultinom.R
\name{dsumuptpois}
\alias{dsumuptpois}
\title{Sum of upper truncated Poisson random variables}
\usage{
dsumuptpois(N, lambda, m)
}
\arguments{
\item{N}{integer (potentially a vector), the value of summed upper truncated Poisson r.v.'s}

\item{lambda}{vector of means for each upper truncted Poisson}

\item{m}{integer vector upper (inclusive) limits}
}
\value{
a vector of probabilities equal in length to \code{N}
}
\description{
Sum of upper truncated Poisson random variables
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
