% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmultinom.R
\name{duptpois}
\alias{duptpois}
\title{Upper truncated Poisson random variables}
\usage{
duptpois(x, lambda, m)
}
\arguments{
\item{x}{vector of integer values at which to calculate probabilities}

\item{lambda}{mean of the upper truncted Poisson}

\item{m}{upper (inclusive) limit}
}
\description{
Upper truncated Poisson random variables
}
\references{
Levin, B. (1981). A representation for multinomial cumulative distribution functions.
The Annals of Statistics, 1123-1126.
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
